DBFS API 429 retries: Fix bug on computing the time until the next retry (#326)
In https://github.com/databricks/databricks-cli/pull/319 we added 429 retries for the DBFS API.

Computing the time until the next retry as `retry_state.idle_for - time_for_last_retry` is not going to work as expected in the case where more than 1 call to functions wrapped with `retry_429` happens (e.g. in the case of `put_file`). The problem is that `time_for_last_retry` is only initialised once per databricks-cli command, while it should be initialised every time a function wrapped with `retry_429` is called, in the first attempt.

We fix this by initialising `time_for_last_retry` if `retry_state.attempt_number == 1`. This only works if the code is single-threaded, which is the case here.

Additional improvements:
* Remove unnecessary retry log, since we already have one.
* Increase max number of retries to 8, so that we have one additional retry with long delay.
* Use click.echo instead of logger, to maintain consistency with the rest of the code.
* Add clarifying note on `retry_state.idle_for`.
* Display seconds with 2 decimal places only.