Add tests to ensure API compatibility when using positional arguments (#507)
This test requires the environment variable DATABRICKS_CLI_MAIN_CHECKOUT to be set to the path of a reference checkout of this repository. It will dynamically load the argspecs from both the reference repository and the current repository and compare them. The test fails if the current repository contains incompatible modifications of function signatures.

In CI we can run this test with DATABRICKS_CLI_MAIN_CHECKOUT pointing to a checkout of the main branch. It will test compatibility of the public API in the PR against the public API in the main branch, and fail on error.